% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Latinobarometro_2023_Costa_Rica_Spss_esp_v1.R
\name{Latinobarometro_2023_Costa_Rica_Spss_esp_v1}
\alias{Latinobarometro_2023_Costa_Rica_Spss_esp_v1}
\title{Latinobarometro 2023 Costa Rica Survey Data}
\format{
A data frame with survey responses from Costa Rican participants.
Variables include demographic information and survey responses.
}
\source{
Latinobarometro 2023. \url{https://www.latinobarometro.org/}
}
\description{
Survey data from the Latinobarometro 2023 study conducted in Costa Rica.
This dataset contains responses from Costa Rican participants on various
social, political, and economic topics.
}
\keyword{datasets}
