% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners.R
\docType{data}
\name{runners}
\alias{runners}
\title{runners}
\format{
A data frame with 80 rows and 3 columns:
\describe{
\item{Time}{numeric, race time in minutes.}
\item{Age}{numeric, runner's age.}
\item{Miles}{numeric, average miles per week in training.}
}
}
\usage{
runners
}
\description{
Hypothetical data gathered from 80 individuals who are all runners.
Three variables have been measured:
\itemize{
\item Time: each runner's most recent 5-kilometer race time, measured in minutes.
\item Age: each runner's age.
\item Miles: the average number of miles per week that each runner runs when in training for a race.
}
}
\keyword{datasets}
